requirements.txt
Flask==3.0.0
PyGithub==2.1.1
plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0

